clients/serializers.py
from django.contrib.auth.models import User
from rest_framework import serializers


class ClientSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Client
        fields = ('name', 'surname', 'birth_Day')
