requirements.txt
xlsxwriter
openpyxl
Django==1.11.1
django-filter==1.0.4
django-filters==0.2.1
djangorestframework==3.6.3
olefile==0.44
Pillow==4.1.1
pkg-resources==0.0.0
pytz==2017.2
